Typed dict implementation
Changes to jitted functions/classes to use the typed dicts provided by the new utility funciton in cfg.py. This is now called in the cosipy_core. This way, the user can interact with the NAMELIST in the familiar form of a python dictionary. But when the run is initilaized, this is split into typed dictionaries compatible with numba.